Move Softplus components to respective modules
Also adds an inplace version
And fixed the jax code to be equivalent to the Aesara implementation
