+++ aesara/tensor/nnet/__init__.py
@@ -43,4 +43,4 @@ from aesara.tensor.nnet.basic import (
     softsign,
 )
 from aesara.tensor.nnet.batchnorm import batch_normalization
-from aesara.tensor.nnet.sigm import hard_sigmoid, softplus, ultra_fast_sigmoid
+from aesara.tensor.nnet.sigm import hard_sigmoid, ultra_fast_sigmoid
